requirements.txt
cloudpickle==2.0.0
cycler==0.11.0
fonttools==4.31.2
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.0
matplotlib==3.5.1
numba==0.55.1
numpy==1.21.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.8.0
shap==0.40.0
six==1.16.0
slicer==0.0.7
threadpoolctl==3.1.0
tqdm==4.63.1
yapf==0.32.0
